Move hashing function to the corresponding *ramp file
