+++ chains/solana/contracts/programs/ccip-router/src/instructions/v1/offramp.rs
@@ -2,7 +2,7 @@ use anchor_lang::prelude::*;
 use solana_program::{instruction::Instruction, program::invoke_signed};
 
 use super::merkle::{calculate_merkle_root, MerkleError};
-use super::messages::{hash_any2solana, ReleaseOrMintInV1, ReleaseOrMintOutV1};
+use super::messages::{ReleaseOrMintInV1, ReleaseOrMintOutV1};
 use super::ocr3base::{ocr3_transmit, ReportContext};
 use super::ocr3impl::{Ocr3ReportForCommit, Ocr3ReportForExecutionReportSingleChain};
 use super::pools::{
@@ -11,13 +11,14 @@ use super::pools::{
 };
 
 use crate::{
-    Any2SolanaMessage, BillingTokenConfigWrapper, CcipRouterError, CommitInput, CommitReport,
-    CommitReportAccepted, CommitReportContext, DestChain, ExecuteReportContext,
-    ExecutionReportSingleChain, ExecutionStateChanged, GasPriceUpdate, GlobalState,
-    MessageExecutionState, OcrPluginType, RampMessageHeader, SkippedAlreadyExecutedMessage,
-    SolanaAccountMeta, SolanaTokenAmount, SourceChain, TimestampedPackedU224, TokenPriceUpdate,
-    UsdPerTokenUpdated, UsdPerUnitGasUpdated, CCIP_RECEIVE_DISCRIMINATOR, DEST_CHAIN_STATE_SEED,
-    EXTERNAL_EXECUTION_CONFIG_SEED, EXTERNAL_TOKEN_POOL_SEED, FEE_BILLING_TOKEN_CONFIG, STATE_SEED,
+    Any2SolanaMessage, Any2SolanaRampMessage, BillingTokenConfigWrapper, CcipRouterError,
+    CommitInput, CommitReport, CommitReportAccepted, CommitReportContext, DestChain,
+    ExecuteReportContext, ExecutionReportSingleChain, ExecutionStateChanged, GasPriceUpdate,
+    GlobalState, MessageExecutionState, OcrPluginType, RampMessageHeader,
+    SkippedAlreadyExecutedMessage, SolanaAccountMeta, SolanaTokenAmount, SourceChain,
+    TimestampedPackedU224, TokenPriceUpdate, UsdPerTokenUpdated, UsdPerUnitGasUpdated,
+    CCIP_RECEIVE_DISCRIMINATOR, DEST_CHAIN_STATE_SEED, EXTERNAL_EXECUTION_CONFIG_SEED,
+    EXTERNAL_TOKEN_POOL_SEED, FEE_BILLING_TOKEN_CONFIG, STATE_SEED,
 };
 
 pub fn commit<'info>(
@@ -613,7 +614,7 @@ pub fn verify_merkle_root(
     execution_report: &ExecutionReportSingleChain,
     on_ramp_address: &[u8],
 ) -> Result<[u8; 32]> {
-    let hashed_leaf = hash_any2solana(&execution_report.message, on_ramp_address);
+    let hashed_leaf = hash(&execution_report.message, on_ramp_address);
     let verified_root: std::result::Result<[u8; 32], MerkleError> =
         calculate_merkle_root(hashed_leaf, execution_report.proofs.clone());
     require!(
@@ -661,4 +662,96 @@ pub fn validate_execution_report<'info>(
     );
 
     Ok(())
+}
+
+fn hash(msg: &Any2SolanaRampMessage, on_ramp_address: &[u8]) -> [u8; 32] {
+    use anchor_lang::solana_program::hash;
+
+    // Calculate vectors size to ensure that the hash is unique
+    let sender_size = [msg.sender.len() as u8];
+    let on_ramp_address_size = [on_ramp_address.len() as u8];
+    let data_size = msg.data.len() as u16; // u16 > maximum transaction size, u8 may have overflow
+
+    // RampMessageHeader struct
+    let header_source_chain_selector = msg.header.source_chain_selector.to_be_bytes();
+    let header_dest_chain_selector = msg.header.dest_chain_selector.to_be_bytes();
+    let header_sequence_number = msg.header.sequence_number.to_be_bytes();
+    let header_nonce = msg.header.nonce.to_be_bytes();
+
+    // Extra Args struct
+    let extra_args_compute_units = msg.extra_args.compute_units.to_be_bytes();
+    let extra_args_accounts_len = [msg.extra_args.accounts.len() as u8];
+    let extra_args_accounts = msg.extra_args.accounts.try_to_vec().unwrap();
+
+    // TODO: Hash token amounts
+
+    // NOTE: calling hash::hashv is orders of magnitude cheaper than using Hasher::hashv
+    // As similar as https://github.com/smartcontractkit/chainlink/blob/develop/contracts/src/v0.8/ccip/offRamp/OffRamp.sol#L402
+    let result = hash::hashv(&[
+        "Any2SolanaMessageHashV1".as_bytes(),
+        &header_source_chain_selector,
+        &header_dest_chain_selector,
+        &on_ramp_address_size,
+        on_ramp_address,
+        &msg.header.message_id,
+        &msg.receiver.to_bytes(),
+        &header_sequence_number,
+        &extra_args_compute_units,
+        &extra_args_accounts_len,
+        &extra_args_accounts,
+        &header_nonce,
+        &sender_size,
+        &msg.sender,
+        &data_size.to_be_bytes(),
+        &msg.data,
+    ]);
+
+    result.to_bytes()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{Any2SolanaRampMessage, SolanaExtraArgs};
+
+    /// Builds a message and hash it, it's compared with a known hash
+    #[test]
+    fn test_hash() {
+        let message = Any2SolanaRampMessage {
+            sender: [
+                1, 2, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                0, 0, 0, 0,
+            ]
+            .to_vec(),
+            receiver: Pubkey::try_from("DS2tt4BX7YwCw7yrDNwbAdnYrxjeCPeGJbHmZEYC8RTb").unwrap(),
+            data: vec![4, 5, 6],
+            header: RampMessageHeader {
+                message_id: [
+                    8, 5, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                    0, 0, 0, 0, 0, 0,
+                ],
+                source_chain_selector: 67,
+                dest_chain_selector: 78,
+                sequence_number: 89,
+                nonce: 90,
+            },
+            token_amounts: [].to_vec(), // TODO: hash token amounts
+            extra_args: SolanaExtraArgs {
+                compute_units: 1000,
+                accounts: vec![SolanaAccountMeta {
+                    pubkey: Pubkey::try_from("DS2tt4BX7YwCw7yrDNwbAdnYrxjeCPeGJbHmZEYC8RTb")
+                        .unwrap(),
+                    is_writable: true,
+                }],
+            },
+        };
+
+        let on_ramp_address = &[1, 2, 3].to_vec();
+        let hash_result = hash(&message, on_ramp_address);
+
+        assert_eq!(
+            "03da97f96c82237d8a8ab0f68d4f7ba02afe188b4a876f348278fbf2226312ed",
+            hex::encode(hash_result)
+        );
+    }
 }
